damg7374/code/frontend/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
entrypoints==0.4
Faker==33.1.0
favicon==0.7.0
fonttools==4.55.2
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
htbuilder==0.7.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
more-itertools==10.5.0
narwhals==1.17.0
neo4j==5.27.0
numpy==2.2.0
openai==1.57.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
prometheus_client==0.21.1
protobuf==5.29.1
pyarrow==18.1.0
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pymdown-extensions==10.12
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
st-annotated-text==4.0.1
st-theme==1.2.3
streamlit==1.41.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-float==0.3.5
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validators==0.34.0
